requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
einops==0.8.1
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
hyperframe==6.1.0
HyperPyYAML==1.2.2
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
julius==0.2.7
PyJWT==2.10.1
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
openai==1.107.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
postgrest==1.1.1
primePy==1.3
propcache==0.3.2
protobuf==6.32.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannote.audio==3.4.0
pycparser==2.23
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pymysql==1.1.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.43
starlette==0.47.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.11.0
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
tqdm==4.67.1
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
googletrans==4.0.0-rc1